Add client

+++ raytracer-server/src/lib/operation.rs
@@ -10,7 +10,7 @@ pub trait RenderStateOperation {
 impl RenderStateOperation for Operation {
     fn apply_to_render_state(&self, state: &mut RenderState) {
         match self {
-            Operation::SetFov(Fov { fov }) => state.fov = *fov,
+            Operation::SetFov(Fov { fov }) => state.fov = *fov as f32,
             Operation::SetResolution(Resolution { width, height }) => {
                 state.width = *width as usize;
                 state.height = *height as usize;
@@ -19,9 +19,9 @@ impl RenderStateOperation for Operation {
                 if direction.is_some() {
                     let dir = direction.clone().unwrap();
                     state.view_dir = Vec3 {
-                        x: dir.x,
-                        y: dir.y,
-                        z: dir.z,
+                        x: dir.x as f32,
+                        y: dir.y as f32,
+                        z: dir.z as f32,
                     }
                 }
             }
@@ -29,9 +29,9 @@ impl RenderStateOperation for Operation {
                 if origin.is_some() {
                     let origin = origin.clone().unwrap();
                     state.origin = Vec3 {
-                        x: origin.x,
-                        y: origin.y,
-                        z: origin.z,
+                        x: origin.x as f32,
+                        y: origin.y as f32,
+                        z: origin.z as f32,
                     }
                 }
             }
